Add cell for the habit creation name field
