+++ Habit-Calendar/Habit-Calendar/Views/Empty State View/pt-BR.lproj/EmptyStateView.strings
