+++ Habit-Calendar/Habit-Calendar/Views/RoundedButton.swift
@@ -48,6 +48,7 @@ import UIKit
 
     override func prepareForInterfaceBuilder() {
         super.prepareForInterfaceBuilder()
+        setup()
         layoutIfNeeded()
     }
